+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Upgrade styled-components@^4.0.0
 - Remove built-in Header in latest-section of index-page
 - Remove text-decoration of links in Footer
+- Replace the `OutboundLink` of `react-ga` with customized `TrackedLink`
 
 ### 5.1.3
 - Add url event label to google analysis click event